refactor(config): change some file location
