--- config.json
@@ -1,7 +1,0 @@
-{
-    "token": "YOUR_TOKEN",
-    "activityType": "PLAYING",
-    "activity": "your music selections",
-    "clientId": "YOUR_CLIENT_ID",
-    "guildId": "YOUR_GUILD_ID,TARGET_SERVER_GUILD_ID",
-}
