+++ helper/deploy-command.js
@@ -1,5 +1,5 @@
 const { REST, Routes } = require('discord.js');
-const { clientId, guildId, token } = require('../config.json');
+const { clientId, guildId, token } = require('../config/config.json');
 const fs = require('node:fs');
 
 const commands = [];
